+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][chg] and this project adheres to
 [Haskell's Package Versioning Policy][pvp]
 
+## Unreleased
+
+  - Add `KafkaEvent` type for subscribing Lambda functions to Kafka
+    topics
+  - Support `aeson ^>=2.0.0.0`
+
 ## `0.4.8` - 2021-05-07
 
   - Add `ToJSON` instances for `ProxyRequest` types, and test